+++ src/main/java/org/folio/search/repository/ConsortiumInstitutionRepository.java
@@ -1,6 +1,7 @@
 package org.folio.search.repository;
 
 import static org.folio.search.model.types.ResourceType.INSTITUTION;
+import static org.folio.search.utils.SearchUtils.ID_FIELD;
 import static org.folio.search.utils.SearchUtils.TENANT_ID_FIELD_NAME;
 import static org.folio.search.utils.SearchUtils.performExceptionalOperation;
 import static org.opensearch.search.sort.SortOrder.ASC;
@@ -36,27 +37,38 @@ public class ConsortiumInstitutionRepository {
 
   public SearchResult<ConsortiumInstitution> fetchInstitutions(String tenantHeader,
                                                                String tenantId,
+                                                               String id,
                                                                Integer limit,
                                                                Integer offset,
                                                                String sortBy,
                                                                SortOrder sortOrder) {
 
-    var sourceBuilder = getSearchSourceBuilder(tenantId, limit, offset, sortBy, sortOrder);
+    var sourceBuilder = getSearchSourceBuilder(tenantId, id, limit, offset, sortBy, sortOrder);
     var response = search(sourceBuilder, tenantHeader);
     return documentConverter.convertToSearchResult(response, ConsortiumInstitution.class);
   }
 
   @NotNull
   private static SearchSourceBuilder getSearchSourceBuilder(String tenantId,
+                                                            String institutionId,
                                                             Integer limit,
                                                             Integer offset,
                                                             String sortBy,
                                                             SortOrder sortOrder) {
     var sourceBuilder = new SearchSourceBuilder();
+    var boolQuery = QueryBuilders.boolQuery();
+
     Optional.ofNullable(tenantId)
-      .ifPresent(id -> sourceBuilder
-        .query(QueryBuilders.boolQuery()
-          .filter(QueryBuilders.termQuery(TENANT_ID_FIELD_NAME, id))));
+      .ifPresent(id -> boolQuery
+          .filter(QueryBuilders.termQuery(TENANT_ID_FIELD_NAME, id)));
+
+    Optional.ofNullable(institutionId)
+        .ifPresent(id -> boolQuery
+            .filter(QueryBuilders.termQuery(ID_FIELD, id)));
+
+    if (boolQuery.hasClauses()) {
+      sourceBuilder.query(boolQuery);
+    }
 
     return sourceBuilder
       .from(offset)